Create a new file in order to add the terminal menu
