+++ src/ui/mod.rs
@@ -1,6 +1,7 @@
 pub mod menu;
 pub mod splash;
 pub mod game;
+pub mod terminal_menu;
 
 use bevy::prelude::*;
